+++ app/about/page.tsx
@@ -1,0 +1,5 @@
+import { CmsPublicPage } from "@/components/cms/cms-public-page";
+
+export default function AboutPage() {
+  return <CmsPublicPage path="/curator" />;
+}